imp: Move tests and fixtures
