+++ packages/platform-ios/src/commands/runIOS/__tests__/findMatchingSimulator.test.ts
@@ -5,8 +5,6 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @format
- * @emails oncall+javascript_foundation
  */
 
 import findMatchingSimulator from '../findMatchingSimulator';
@@ -71,33 +69,33 @@ describe('findMatchingSimulator', () => {
             'iOS 12.1': [
               {
                 state: 'Shutdown',
-                isAvailable: 'YES',
+                isAvailable: true,
                 name: 'iPhone 6s',
                 udid: 'D0F29BE7-CC3C-4976-888D-C739B4F50508',
               },
               {
                 state: 'Shutdown',
-                isAvailable: 'YES',
+                isAvailable: true,
                 name: 'iPhone 6',
                 udid: 'BA0D93BD-07E6-4182-9B0A-F60A2474139C',
               },
               {
                 state: 'Shutdown',
-                isAvailable: 'YES',
+                isAvailable: true,
                 name: 'iPhone XS Max',
                 udid: 'B9B5E161-416B-43C4-A78F-729CB96CC8C6',
                 availabilityError: '',
               },
               {
                 state: 'Shutdown',
-                isAvailable: 'YES',
+                isAvailable: true,
                 name: 'iPad Air',
                 udid: '1CCBBF8B-5773-4EA6-BD6F-C308C87A1ADB',
                 availabilityError: '',
               },
               {
                 state: 'Shutdown',
-                isAvailable: 'YES',
+                isAvailable: true,
                 name: 'iPad (5th generation)',
                 udid: '9564ABEE-9EC2-4B4A-B443-D3710929A45A',
                 availabilityError: '',
@@ -157,12 +155,6 @@ describe('findMatchingSimulator', () => {
         'iPhone 6',
       ),
     ).toEqual(null);
-  });
-
-  it('should return null if an odd input', () => {
-    expect(findMatchingSimulator('random string input', 'iPhone 6')).toEqual(
-      null,
-    );
   });
 
   it('should return the first simulator in list if none is defined', () => {
