+++ packages/platform-ios/src/config/__tests__/findPodspec.test.ts
@@ -4,7 +4,6 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @flow
  */
 
 import findPodspec from '../findPodspec';
