+++ packages/platform-ios/src/link-pods/__tests__/findMarkedLinesInPodfile.test.ts
@@ -4,8 +4,6 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @format
- * @emails oncall+javascript_foundation
  */
 
 import readPodfile from '../readPodfile';
@@ -14,7 +12,6 @@ import findMarkedLinesInPodfile from '../findMarkedLinesInPodfile';
 const path = require('path');
 
 const PODFILES_PATH = path.join(__dirname, '../__fixtures__');
-const LINE_AFTER_TARGET_IN_TEST_PODFILE = 4;
 
 describe('pods::findMarkedLinesInPodfile', () => {
   it('returns empty array if file is not Podfile', () => {
@@ -24,9 +21,7 @@ describe('pods::findMarkedLinesInPodfile', () => {
 
   it('returns empty array for Simple Podfile', () => {
     const podfile = readPodfile(path.join(PODFILES_PATH, 'PodfileSimple'));
-    expect(
-      findMarkedLinesInPodfile(podfile, LINE_AFTER_TARGET_IN_TEST_PODFILE),
-    ).toEqual([]);
+    expect(findMarkedLinesInPodfile(podfile)).toEqual([]);
   });
 
   it('returns correct line numbers for Podfile with marker', () => {
@@ -35,8 +30,6 @@ describe('pods::findMarkedLinesInPodfile', () => {
       {line: 18, indentation: 2},
       {line: 31, indentation: 4},
     ];
-    expect(
-      findMarkedLinesInPodfile(podfile, LINE_AFTER_TARGET_IN_TEST_PODFILE),
-    ).toEqual(expectedObject);
+    expect(findMarkedLinesInPodfile(podfile)).toEqual(expectedObject);
   });
 });
