+++ packages/platform-ios/src/link/__tests__/getPlistPath.test.ts
@@ -4,11 +4,9 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @format
- * @emails oncall+javascript_foundation
  */
 
-import getBuildProperty from '../getBuildProperty';
+import getPlistPath from '../getPlistPath';
 
 const xcode = require('xcode');
 const path = require('path');
@@ -17,13 +15,13 @@ const project = xcode.project(
   path.join(__dirname, '../__fixtures__/project.pbxproj'),
 );
 
-describe('ios::getBuildProperty', () => {
+describe('ios::getPlistPath', () => {
   beforeEach(() => {
     project.parseSync();
   });
 
-  it('should return build property from main target', () => {
-    const plistPath = getBuildProperty(project, 'INFOPLIST_FILE');
-    expect(plistPath).toEqual('Basic/Info.plist');
+  it('should return path without Xcode $(SRCROOT)', () => {
+    const plistPath = getPlistPath(project, '/');
+    expect(plistPath).toBe(path.normalize('/Basic/Info.plist'));
   });
 });
