+++ packages/platform-ios/src/link/__tests__/isInstalled.test.ts
@@ -4,8 +4,6 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @format
- * @emails oncall+javascript_foundation
  */
 
 import isInstalled from '../isInstalled';
@@ -20,11 +18,13 @@ const baseProjectConfig = {
 describe('ios::isInstalled', () => {
   it('should return true when .xcodeproj in Libraries', () => {
     const dependencyConfig = {projectName: 'React.xcodeproj'};
+    // @ts-ignore FIXME: Improve types
     expect(isInstalled(baseProjectConfig, dependencyConfig)).toBeTruthy();
   });
 
   it('should return false when .xcodeproj not in Libraries', () => {
     const dependencyConfig = {projectName: 'Missing.xcodeproj'};
+    // @ts-ignore FIXME: Improve types
     expect(isInstalled(baseProjectConfig, dependencyConfig)).toBeFalsy();
   });
 
@@ -33,6 +33,7 @@ describe('ios::isInstalled', () => {
     const projectConfig = Object.assign({}, baseProjectConfig, {
       libraryFolder: 'Missing',
     });
+    // @ts-ignore FIXME: Improve types
     expect(isInstalled(projectConfig, dependencyConfig)).toBeFalsy();
   });
 });
