+++ packages/platform-ios/src/link/__tests__/removeProjectFromLibraries.test.ts
@@ -4,8 +4,6 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @format
- * @emails oncall+javascript_foundation
  */
 
 import addProjectToLibraries from '../addProjectToLibraries';
