+++ packages/platform-ios/src/types.ts
@@ -1,8 +1,9 @@
 export interface Device {
-  availability: string;
-  state: string;
-  isAvailable: boolean;
+  availability?: string;
+  state?: string;
+  isAvailable?: boolean;
   name: string;
   udid: string;
-  availabilityError: string;
+  version?: string;
+  availabilityError?: string;
 }
